Add docs for `UnspannedTable#nokogiri_node` method
We want to provide methods for getting to the `Nokogiri::Node` instance
as well as just a plain string to avoid having to reparse tables after
they've been unspanned.

I've removed the example output now that I'm not using a `puts`. I'll
add another commit after this one to make it clearer what the expected
input and output is.